reorganize to prevent cyclic dependencies
