+++ pkg/config/config.go
@@ -4,18 +4,16 @@ import (
 	"bytes"
 	"flag"
 	"fmt"
-	"io"
 	"os"
 	"strings"
 	"testing"
 	"unicode"
-	"unicode/utf8"
 
-	"github.com/dimchansky/utfbom"
 	"github.com/drone/envsubst/v2"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/grafana/agent/pkg/build"
+	"github.com/grafana/agent/pkg/config/encoder"
 	"github.com/grafana/agent/pkg/config/features"
 	"github.com/grafana/agent/pkg/config/instrumentation"
 	"github.com/grafana/agent/pkg/logs"
@@ -25,9 +23,6 @@ import (
 	"github.com/grafana/agent/pkg/util"
 	"github.com/prometheus/common/config"
 	"github.com/stretchr/testify/require"
-	"golang.org/x/text/encoding"
-	uni "golang.org/x/text/encoding/unicode"
-	"golang.org/x/text/encoding/unicode/utf32"
 	"gopkg.in/yaml.v2"
 )
 
@@ -344,7 +339,7 @@ func LoadRemote(url string, expandEnvVars bool, c *Config) error {
 }
 
 func performEnvVarExpansion(buf []byte, expandEnvVars bool) ([]byte, error) {
-	utf8Buf, err := EnsureUTF8(buf, false)
+	utf8Buf, err := encoder.EnsureUTF8(buf, false)
 	if err != nil {
 		return nil, err
 	}
@@ -369,43 +364,6 @@ func LoadBytes(buf []byte, expandEnvVars bool, c *Config) error {
 	}
 	// Unmarshal yaml config
 	return yaml.UnmarshalStrict(expandedBuf, c)
-}
-
-// EnsureUTF8 will convert from the most common encodings to UTF8.
-// If useStrictUTF8 is enabled then if the file is not already utf8 then an error will be returned.
-func EnsureUTF8(config []byte, useStrictUTF8 bool) ([]byte, error) {
-	buffer := bytes.NewBuffer(config)
-	src, enc := utfbom.Skip(buffer)
-	var converted []byte
-	skippedBytes, err := io.ReadAll(src)
-	if err != nil {
-		return nil, err
-	}
-	var encoder encoding.Encoding
-	switch enc {
-	case utfbom.UTF16BigEndian:
-		encoder = uni.UTF16(uni.BigEndian, uni.IgnoreBOM)
-	case utfbom.UTF16LittleEndian:
-		encoder = uni.UTF16(uni.LittleEndian, uni.IgnoreBOM)
-	case utfbom.UTF32BigEndian:
-		encoder = utf32.UTF32(utf32.BigEndian, utf32.IgnoreBOM)
-	case utfbom.UTF32LittleEndian:
-		encoder = utf32.UTF32(utf32.LittleEndian, utf32.IgnoreBOM)
-	case utfbom.UTF8: // This only checks utf8 bom
-		return config, nil
-	default:
-		// If its utf8 valid then return.
-		if utf8.Valid(config) {
-			return config, nil
-		}
-		return nil, fmt.Errorf("unknown encoding for config")
-	}
-	if useStrictUTF8 {
-		return nil, fmt.Errorf("configuration is encoded with %s but must be utf8", enc.String())
-	}
-	decoder := encoder.NewDecoder()
-	converted, err = decoder.Bytes(skippedBytes)
-	return converted, err
 }
 
 // getenv is a wrapper around os.Getenv that ignores patterns that are numeric
