+++ pkg/config/encoder/test_encoding_utf32be.txt
