+++ pkg/config/encoder/test_encoding_utf8.txt
